+++ addon/services/log.js
@@ -81,6 +81,32 @@ export default Ember.Service.extend(Ember.Evented, {
   enabled: false,
 
   /**
+    The name of a model that represents log entity.
+
+    @property applicationLogModelName
+    @type String
+    @default 'i-i-s-caseberry-logging-objects-application-log'
+    @example
+    ```
+    // Log service 'applicationLogModelName' setting could be also defined through application config/environment.js
+    module.exports = function(environment) {
+      var ENV = {
+        ...
+        APP: {
+          ...
+          log: {
+            enabled: true,
+            applicationLogModelName: 'custom-application-log'
+          }
+          ...
+        }
+        ...
+    };
+    ```
+  */
+  applicationLogModelName: 'i-i-s-caseberry-logging-objects-application-log',
+
+  /**
     Flag: indicates whether log service will store error messages to application log or not.
 
     @property storeErrorMessages
@@ -298,6 +324,8 @@ export default Ember.Service.extend(Ember.Evented, {
     let _this = this;
     let originalMethodsCache = Ember.A();
 
+    this.initProperties();
+
     let originalEmberLoggerError = Ember.Logger.error;
     originalMethodsCache.pushObject({
       methodOwner: Ember.Logger,
@@ -397,6 +425,28 @@ export default Ember.Service.extend(Ember.Evented, {
   },
 
   /**
+   * Initializes properties of a log service.
+   */
+  initProperties() {
+    const config = getOwner(this).resolveRegistration('config:environment');
+    const logConfiguration = config.APP.log;
+
+    this.set('enabled', typeof logConfiguration.enabled === 'boolean' && logConfiguration.enabled);
+    this.set('storeErrorMessages', typeof logConfiguration.storeErrorMessages === 'boolean' && logConfiguration.storeErrorMessages);
+    this.set('storeWarnMessages', typeof logConfiguration.storeWarnMessages === 'boolean' && logConfiguration.storeWarnMessages);
+    this.set('storeLogMessages', typeof logConfiguration.storeLogMessages === 'boolean' && logConfiguration.storeLogMessages);
+    this.set('storeInfoMessages', typeof logConfiguration.storeInfoMessages === 'boolean' && logConfiguration.storeInfoMessages);
+    this.set('storeDebugMessages', typeof logConfiguration.storeDebugMessages === 'boolean' && logConfiguration.storeDebugMessages);
+    this.set('storeDeprecationMessages', typeof logConfiguration.storeDeprecationMessages === 'boolean' && logConfiguration.storeDeprecationMessages);
+    this.set('storePromiseErrors', typeof logConfiguration.storePromiseErrors === 'boolean' && logConfiguration.storePromiseErrors);
+    this.set('showPromiseErrors', typeof logConfiguration.showPromiseErrors === 'boolean' && logConfiguration.showPromiseErrors);
+
+    if (typeof logConfiguration.applicationLogModelName === 'string') {
+      this.set('applicationLogModelName', logConfiguration.applicationLogModelName);
+    }
+  },
+
+  /**
     Destroys log service.
   */
   willDestroy() {
@@ -457,7 +507,7 @@ export default Ember.Service.extend(Ember.Evented, {
       formattedMessage: formattedMessage
     };
 
-    let applicationLogModelName = 'i-i-s-caseberry-logging-objects-application-log';
+    const applicationLogModelName = this.get('applicationLogModelName');
     let store = this.get('store');
 
     // Break if message already exists in store (to avoid infinit loop when message is generated while saving itself).
@@ -479,7 +529,7 @@ export default Ember.Service.extend(Ember.Evented, {
   },
 
   _triggerEvent(eventName, applicationLogModel) {
-    Ember.assert('Logger Error: event name should be a string', Ember.typeOf(eventName) === 'string');
+    Ember.assert('Logger Error: event name should be a string', typeof eventName === 'string');
     let eventNameToTrigger = eventName.toLowerCase();
     this.trigger(eventNameToTrigger, applicationLogModel);
   },
@@ -489,7 +539,7 @@ export default Ember.Service.extend(Ember.Evented, {
       return;
     }
 
-    if (Ember.typeOf(error) === 'string') {
+    if (typeof error === 'string') {
       error = new Error(error);
     }